% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC.diffExpression.multiVarCox.R
\name{ICGC.diffExpression.multiVarCox}
\alias{ICGC.diffExpression.multiVarCox}
\title{To do Cox regression of genes}
\usage{
ICGC.diffExpression.multiVarCox(getData.data, diffExpr.data,
  diffExpr.Lasso.data)
}
\arguments{
\item{getData.data}{dataset from command ICGC.getData}

\item{diffExpr.data}{dataset from command ICGC.diffExpression}

\item{diffExpr.Lasso.data}{dataset from command ICGC.diffExpression_Lasso}
}
\value{
six lists: coxphdata, coxsum, coxsumHR, coxphforest, RiskLevel.coxph and RiskLevel.forest
}
\description{
To do Cox regression of genes
}
\examples{
diffExpr.multiVarCox.data=ICGC.diffExpression.multiVarCox(getData.data,diffExpr.data,diffExpr.Lasso.data)
}
