% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC.diffExpression.R
\name{ICGC.diffExpression}
\alias{ICGC.diffExpression}
\title{To do different expression matrix and genes}
\usage{
ICGC.diffExpression(getData.data)
}
\arguments{
\item{getData.data}{dataset from ICGC.getData command, must include donor and exp_seq}
}
\value{
two lists, ExpressionMatrix and diffGene
}
\description{
To do different expression matrix and genes
}
\examples{
result=ICGC.diffExpression(getData.data)
}
