% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC.diffExpression_heatmap.R
\name{ICGC.diffExpression_heatmap}
\alias{ICGC.diffExpression_heatmap}
\title{To plot heatmap for diff expression genes}
\usage{
ICGC.diffExpression_heatmap(diffExpr.data)
}
\arguments{
\item{diffExpr.data}{two list ExpressionMatrix and diffGene from ICGC.diffExpression command,}
}
\value{
a plot
}
\description{
To plot heatmap for diff expression genes
}
\examples{
ICGC.diffExpression_heatmap(diffExpr.data)
}
