% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC.diffExpression_kmAnalysis.R
\name{ICGC.diffExpression_kmAnalysis}
\alias{ICGC.diffExpression_kmAnalysis}
\title{To do KM analysis for ICGC diffExpression data}
\usage{
ICGC.diffExpression_kmAnalysis(getData.data, diffExpr.data,
  p.filter = 0.05)
}
\arguments{
\item{getData.data}{From command ICGC.getData, must include donor data}

\item{diffExpr.data}{From command ICGC.diffExpression}

\item{p.filter}{the p value of significant difference, EX:0.05, defaulted.}
}
\value{
a dataframe and survival curves end with tiff
}
\description{
To do KM analysis for ICGC diffExpression data
}
\examples{
ICGC.diffExpression_kmAnalysis(getData.data,diffExpr.data)
}
