% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC.diffExpression_nomogram.R
\name{ICGC.diffExpression_nomogram}
\alias{ICGC.diffExpression_nomogram}
\title{To plot nomogram for cox regression}
\usage{
ICGC.diffExpression_nomogram()
}
\value{
a picturn
}
\description{
To plot nomogram for cox regression
}
\examples{
ICGC.diffExpression_nomogram()
}
