% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC.diffExpression_singleGene.kmAnalysis.R
\name{ICGC.diffExpression_singleGene.kmAnalysis}
\alias{ICGC.diffExpression_singleGene.kmAnalysis}
\title{To do KM analysis for a single gene in ICGC data}
\usage{
ICGC.diffExpression_singleGene.kmAnalysis(getData.data, diffExpr.data,
  Genes)
}
\arguments{
\item{getData.data}{From command ICGC.getData, must include donor data}

\item{diffExpr.data}{From command ICGC.diffExpression}

\item{Genes}{genes that you want to plot km curves}
}
\value{
a survival curve
}
\description{
To do KM analysis for a single gene in ICGC data
}
\examples{
ICGC.diffExpression_singleGene.kmAnalysis(getData.data,diffExpr.data,"TP53")
}
