% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC.diffExpression_singleGene.ROC.R
\name{ICGC.diffExpression_singleGene.ROC}
\alias{ICGC.diffExpression_singleGene.ROC}
\title{To do ROC analysis for a single gene in ICGC data}
\usage{
ICGC.diffExpression_singleGene.ROC(getData.data, diffExpr.data, Genes,
  plotROC = TRUE)
}
\arguments{
\item{getData.data}{From command ICGC.getData, must include donor data}

\item{diffExpr.data}{From command ICGC.diffExpression}

\item{Genes}{genes that you want to plot km curves}

\item{plotROC}{a logical param. whether to plot roc, TRUE as defaulted}
}
\value{
a ROC curve
}
\description{
To do ROC analysis for a single gene in ICGC data
}
\examples{
ICGC.diffExpression_singleGene.ROC(getData.data,diffExpr.data,"TP53")
}
