% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC.diffExpression_uniCox.R
\name{ICGC.diffExpression_univarCox}
\alias{ICGC.diffExpression_univarCox}
\title{To do KM univariable Cox regression for ICGC diffExpression data}
\usage{
ICGC.diffExpression_univarCox(getData.data, diffExpr.data,
  GeneType = "c")
}
\arguments{
\item{getData.data}{From command ICGC.getData, must include donor data}

\item{diffExpr.data}{From command ICGC.diffExpression}

\item{GeneType}{Data type of gene. b as binary. c as continuous, defaulted}
}
\value{
a dataframe and a plot
}
\description{
To do KM univariable Cox regression for ICGC diffExpression data
}
\examples{
ICGC.diffExpression_univarCox(getData.data,diffExpr.data)
}
