% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGC.diffExpression_Volcano.R
\name{ICGC.diffExpression_Volcano}
\alias{ICGC.diffExpression_Volcano}
\title{To plot Volcano for diff genes}
\usage{
ICGC.diffExpression_Volcano(diffExpr.data, log2fc.filter,
  minus.log10fdr.filter)
}
\arguments{
\item{diffExpr.data}{dataset from ICGC.diffExpression command}

\item{log2fc.filter}{log2fc cutoff value}

\item{minus.log10fdr.filter}{-log10(fdr) cutoff value}
}
\value{
two lists: VolcanoData and VolcanoPlot.
}
\description{
To plot Volcano for diff genes
}
\examples{
icgc.Volcano(diffExpr.data,2,10)
}
